Moved browser enum to browsermanager package.
