+++ src/main/java/io/swaglabs/portal/qa/browsermanager/MsEdgeBrowser.java
@@ -4,7 +4,6 @@ import com.microsoft.playwright.Browser;
 import com.microsoft.playwright.BrowserContext;
 import com.microsoft.playwright.BrowserType;
 import com.microsoft.playwright.Playwright;
-import io.swaglabs.portal.qa.constants.BrowserName;
 import io.swaglabs.portal.qa.constants.WebPortalConstants;
 
 import java.util.List;
@@ -15,7 +14,7 @@ public class MsEdgeBrowser implements IBrowser {
         return playwright.chromium().launch(new BrowserType.LaunchOptions()
                         .setHeadless(isHeadless)
                         .setArgs(isHeadless
-                                ? List.of(WebPortalConstants.NEW_HEADLESS, WebPortalConstants.MAXIMIZE_WINDOW)
+                                ? List.of("--headless=new", WebPortalConstants.MAXIMIZE_WINDOW)
                                 : List.of(WebPortalConstants.MAXIMIZE_WINDOW))
                         .setChannel(BrowserName.MS_EDGE.getBrowserType()))
                 .newContext(new Browser.NewContextOptions().setViewportSize(null));
